toolchain: cmake: Off-tree toolchains always gets other.h included
To ensure that an off-tree toolchain always gets the toolchain/other.h
header included, such that it can include the correct header for the
toolchain via the other.h, the define __TOOLCHAIN_CUSTOM__ will be set
when ever TOOLCHAIN_ROOT is specified. As TOOLCHAIN_ROOT is only
specified externally for off-tree toolchains, this has no impact on
in-tree toolchains and their functionality.

Fixes zephyrproject-rtos#36117

Signed-off-by: Danny Oerndrup <daor@demant.com>